Move upgrade logic to v4.13.0 as this is the new target version for this feature
