+++ src/main/java/org/dependencytrack/upgrade/UpgradeItems.java
@@ -40,6 +40,7 @@ class UpgradeItems {
         UPGRADE_ITEMS.add(org.dependencytrack.upgrade.v4100.v4100Updater.class);
         UPGRADE_ITEMS.add(org.dependencytrack.upgrade.v4110.v4110Updater.class);
         UPGRADE_ITEMS.add(org.dependencytrack.upgrade.v4120.v4120Updater.class);
+        UPGRADE_ITEMS.add(org.dependencytrack.upgrade.v4130.v4130Updater.class);
     }
 
     static List<Class<? extends UpgradeItem>> getUpgradeItems() {
