+++ src/main/java/org/dependencytrack/upgrade/v4130/v4130Updater.java
@@ -1,0 +1,48 @@
+/*
+ * This file is part of Dependency-Track.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright (c) OWASP Foundation. All Rights Reserved.
+ */
+package org.dependencytrack.upgrade.v4130;
+
+import alpine.common.logging.Logger;
+import alpine.persistence.AlpineQueryManager;
+import alpine.server.upgrade.AbstractUpgradeItem;
+
+import java.sql.Connection;
+import java.sql.Statement;
+
+public class v4130Updater extends AbstractUpgradeItem {
+
+    private static final Logger LOGGER = Logger.getLogger(v4130Updater.class);
+
+    @Override
+    public String getSchemaVersion() {
+        return "4.13.0";
+    }
+
+    @Override
+    public void executeUpgrade(final AlpineQueryManager qm, final Connection connection) throws Exception {
+        setProjectCollectionLogicDefault(connection);
+    }
+    private static void setProjectCollectionLogicDefault(final Connection connection) throws Exception {
+        LOGGER.info("Setting collection logic of all projects to default NONE");
+        try (final Statement stmt = connection.createStatement()) {
+            stmt.execute("UPDATE \"PROJECT\" SET \"COLLECTION_LOGIC\" = 'NONE'");
+        }
+    }
+
+}
